# api/openapi/mujeeb24-dashboard-v1.generated.yaml
components:
  schemas:
...
    AuthOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/AuthOutputBody.json
          format: uri
          readOnly: true
          type: string
        data:
          $ref: "#/components/schemas/AuthResponse"
        request_id:
          type: string
      required:
        - data
        - request_id
      type: object
    AuthResponse:
      additionalProperties: false
      properties:
        access_token:
          type: string
        expires_at:
          type: string
        principal:
          $ref: "#/components/schemas/Principal"
        token_type:
          type: string
      required:
        - access_token
        - token_type
        - expires_at
        - principal
      type: object
...
    BeginChannelConnectionRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/BeginChannelConnectionRequest.json
          format: uri
          readOnly: true
          type: string
        channel:
          type: string
        display_name:
          type: string
        provider:
          type: string
      required:
        - provider
        - channel
      type: object
    Business:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          type: string
        default_currency:
          type: string
        id:
          type: string
        locale:
          type: string
        name:
          type: string
        resource_version:
          type: string
        slug:
          type: string
        status:
          type: string
        timezone:
          type: string
        updated_at:
          format: date-time
          type: string
        vertical_type:
          type: string
      required:
        - id
        - name
        - slug
        - status
        - vertical_type
        - timezone
        - default_currency
        - locale
        - created_at
        - updated_at
        - resource_version
      type: object
    BusinessMembership:
      additionalProperties: false
      properties:
        business:
          $ref: "#/components/schemas/Business"
        permissions:
          items:
            type: string
          nullable: true
          type: array
        role:
          type: string
      required:
        - business
        - role
      type: object
...
    Capability:
      additionalProperties: false
      properties:
        checked_at:
          format: date-time
          type: string
        enabled:
          type: boolean
        evidence_source:
          type: string
        name:
          type: string
      required:
        - name
        - enabled
        - checked_at
      type: object
...
    ChannelConnection:
      additionalProperties: false
      properties:
        business_id:
          type: string
        capabilities:
          items:
            $ref: "#/components/schemas/Capability"
          nullable: true
          type: array
        channel:
          type: string
        created_at:
          format: date-time
          type: string
        external_account_reference:
          type: string
        id:
          type: string
        provider:
          type: string
        resource_version:
          type: string
        status:
          type: string
        updated_at:
          format: date-time
          type: string
      required:
        - id
        - business_id
        - provider
        - channel
        - status
        - capabilities
        - created_at
        - updated_at
        - resource_version
      type: object
    ChannelProvisioning:
      additionalProperties: false
      properties:
        authorization_url:
          type: string
        business_id:
          type: string
        channel:
          type: string
        created_at:
          format: date-time
          type: string
        provider:
          type: string
        provisioning_id:
          type: string
        status:
          type: string
        updated_at:
          format: date-time
          type: string
      required:
        - provisioning_id
        - business_id
        - provider
        - channel
        - status
      type: object
...
    Cookie:
      additionalProperties: false
      properties:
        Domain:
          type: string
        Expires:
          format: date-time
          type: string
        HttpOnly:
          type: boolean
        MaxAge:
          format: int64
          type: integer
        Name:
          type: string
        Path:
          type: string
        Raw:
          type: string
        RawExpires:
          type: string
        SameSite:
          format: int64
          type: integer
        Secure:
          type: boolean
        Unparsed:
          items:
            type: string
          nullable: true
          type: array
        Value:
          type: string
      required:
        - Name
        - Value
        - Path
        - Domain
        - Expires
        - RawExpires
        - MaxAge
        - Secure
        - HttpOnly
        - SameSite
        - Raw
        - Unparsed
      type: object
...
    ErrorBody:
      additionalProperties: false
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        fields:
          additionalProperties:
            type: string
          type: object
        message:
          type: string
        retryable:
          type: boolean
      required:
        - code
        - message
        - retryable
      type: object
...
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        request_id:
          type: string
      required:
        - error
        - request_id
      type: object
...
    ListBusinessMembershipBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/ListBusinessMembershipBody.json
          format: uri
          readOnly: true
          type: string
        data:
          items:
            $ref: "#/components/schemas/BusinessMembership"
          nullable: true
          type: array
        pagination:
          $ref: "#/components/schemas/Page"
        request_id:
          type: string
      required:
        - data
        - pagination
        - request_id
      type: object
...
    ListChannelConnectionBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/ListChannelConnectionBody.json
          format: uri
          readOnly: true
          type: string
        data:
          items:
            $ref: "#/components/schemas/ChannelConnection"
          nullable: true
          type: array
        pagination:
          $ref: "#/components/schemas/Page"
        request_id:
          type: string
      required:
        - data
        - pagination
        - request_id
      type: object
...
    LoginInputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/LoginInputBody.json
          format: uri
          readOnly: true
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
      type: object
...
    Page:
      additionalProperties: false
      properties:
        has_more:
          type: boolean
        next_cursor:
          type: string
      required:
        - has_more
      type: object
    Principal:
      additionalProperties: false
      properties:
        display_name:
          type: string
        email:
          type: string
        principal_id:
          type: string
      required:
        - principal_id
        - display_name
      type: object
...
    SingleChannelProvisioningBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/SingleChannelProvisioningBody.json
          format: uri
          readOnly: true
          type: string
        data:
          $ref: "#/components/schemas/ChannelProvisioning"
        request_id:
          type: string
      required:
        - data
        - request_id
      type: object
...
    SinglePrincipalBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/SinglePrincipalBody.json
          format: uri
          readOnly: true
          type: string
        data:
          $ref: "#/components/schemas/Principal"
        request_id:
          type: string
      required:
        - data
        - request_id
      type: object
...
    WebhookAccepted:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          example: /api/v1/schemas/WebhookAccepted.json
          format: uri
          readOnly: true
          type: string
        accepted:
          type: boolean
        request_id:
          type: string
      required:
        - accepted
        - request_id
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: EdDSA/Ed25519 JWT access token
      scheme: bearer
      type: http
    refreshCookie:
      description: HttpOnly opaque refresh session cookie
      in: cookie
      name: mujeeb_refresh
      type: apiKey
info:
  title: Mujeeb 24 Dashboard API
  version: 1.0.0
openapi: 3.0.3
paths:
  /auth/login:
    post:
      operationId: authenticatePrincipal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInputBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOutputBody"
          description: OK
          headers:
            Set-Cookie:
              schema:
                $ref: "#/components/schemas/Cookie"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Too Many Requests
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security: []
      summary: Authenticate and issue JWT access token
      tags:
        - Auth
  /auth/logout:
    post:
      operationId: revokeRefreshSession
      parameters:
        - in: header
          name: Cookie
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: Revoke refresh session
      tags:
        - Auth
  /auth/refresh:
    post:
      operationId: rotateRefreshSession
      parameters:
        - in: header
          name: Cookie
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthOutputBody"
          description: OK
          headers:
            Set-Cookie:
              schema:
                $ref: "#/components/schemas/Cookie"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Too Many Requests
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - refreshCookie: []
      summary: Rotate refresh session
      tags:
        - Auth
...
  /businesses/{business_id}/channel-connections:
    get:
      operationId: listChannelConnections
      parameters:
        - in: path
          name: business_id
          required: true
          schema:
            format: uuid
            type: string
        - explode: false
          in: query
          name: limit
          schema:
            default: 25
            format: int64
            maximum: 100
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: cursor
          schema:
            type: string
        - explode: false
          in: query
          name: status
          schema:
            type: string
        - explode: false
          in: query
          name: channel
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChannelConnectionBody"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: List channel connections
      tags:
        - Connections
    post:
      operationId: beginChannelConnection
      parameters:
        - in: path
          name: business_id
          required: true
          schema:
            format: uuid
            type: string
        - in: header
          name: Idempotency-Key
          schema:
            type: string
        - in: header
          name: If-Match
          schema:
            type: string
        - in: header
          name: X-Request-ID
          schema:
            type: string
        - in: header
          name: X-Correlation-ID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeginChannelConnectionRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleChannelProvisioningBody"
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: Begin channel provisioning
      tags:
        - Connections
...
  /me:
    get:
      operationId: getCurrentPrincipal
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SinglePrincipalBody"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: Get the authenticated principal
      tags:
        - Me
  /me/businesses:
    get:
      operationId: listAccessibleBusinesses
      parameters:
        - explode: false
          in: query
          name: limit
          schema:
            default: 25
            format: int64
            maximum: 100
            minimum: 1
            type: integer
        - explode: false
          in: query
          name: cursor
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBusinessMembershipBody"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security:
        - bearerAuth: []
      summary: List businesses accessible to the principal
      tags:
        - Me
...
  /webhooks/socialapi/{route_key}:
    post:
      operationId: ingestSocialAPIWebhook
      parameters:
        - in: path
          name: route_key
          required: true
          schema:
            type: string
        - in: header
          name: X-Provider-Signature
          schema:
            type: string
        - in: header
          name: X-Provider-Timestamp
          schema:
            type: string
        - in: header
          name: X-Request-ID
          schema:
            type: string
        - in: header
          name: X-SocialAPI-Signature
          schema:
            type: string
        - in: header
          name: X-SocialAPI-Signature-V2
          schema:
            type: string
        - in: header
          name: X-SocialAPI-Timestamp
          schema:
            type: string
        - in: header
          name: X-SocialAPI-Delivery
          schema:
            type: string
        - in: header
          name: X-SocialAPI-Event
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema: {}
          application/octet-stream:
            schema:
              contentMediaType: application/octet-stream
              format: binary
              type: string
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAccepted"
          description: Accepted
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unauthorized
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Request Entity Too Large
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Internal Server Error
      security: []
      summary: Ingest SocialAPI webhook
      tags:
        - Webhooks
servers:
  - description: Mujeeb 24 API V1
    url: /api/v1